Refac: iMCache.Repo 에서 캐시 백업 분리
